+++ src/hooks/wallet/useBalance.ts
@@ -1,4 +1,5 @@
-import { PEANUT_WALLET_TOKEN, peanutPublicClient } from '@/constants/zerodev.consts'
+import { PEANUT_WALLET_TOKEN } from '@/constants/zerodev.consts'
+import { peanutPublicClient } from '@/app/actions/clients'
 import { useQuery } from '@tanstack/react-query'
 import { erc20Abi } from 'viem'
 import type { Address } from 'viem'
